requirements.txt
streamlit==1.35.0
pandas==2.2.2
plotly==5.22.0
fpdf==1.7.2
pyrebase4==4.5.0
requests==2.31.0
urllib3==2.2.1
